requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
garminconnect==0.2.19
garth==0.4.46
idna==3.7
lxml==5.2.2
oauthlib==3.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
withings-sync==4.2.4
